backend/requirements-base.txt
# Core ML/AI frameworks (heavyweight - ~500MB+)
transformers>=4.40.0,<4.47.0
sentence-transformers>=3.3.0
accelerate>=1.2.0
tokenizers>=0.19.0,<0.21.0
huggingface-hub>=0.26.0

# Vector databases
pgvector==0.2.4

# Document processing
pypdf==5.6.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-pptx>=0.6.21

# Database drivers
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic>=1.12.0

# Scientific computing
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=1.5.0